Angular and route fixes
In routes: userName, ID, and Provider are now passed into the view.
This will allow all 3 of those to be checked against the poll’s owner
before allowing deletion of a poll.

On Angular, if a user returns to a poll they already visited (and
Angular already rendered), the JS for drawing D3 will use those
existing rendered elements.

Previously, returning to a poll already visited/rendered would result
in the JS inserting in the chart & legend again.  This would cause
multiple copies of the same chart/legend showing up at once.